issue #300 resolved

+++ sfcs_app/app/cutting/controllers/lay_plan_preparation/order_cut_form2.php
@@ -68,7 +68,7 @@ function verify_con(e){
 function percent_cal()
 {
     
-	var x=parseInt(document.input.cuttable_percent.value);
+	var x=parseFloat(document.input.cuttable_percent.value);
 	if(isNaN(x)){
 		// alert();
 		var x = 0;
